backend/requirements.txt
django==3.1.7
djangorestframework==3.11
djangorestframework-simplejwt==5.2.2
pytest==8.3.3
pytest-django==4.9.0
drf-yasg==1.21.8